DepCard: Add a dep-card component (currently just the collapsed view)
Use dy instead of the old max/min so we can use the y centerline
directly.

Also switch the SVG over to a fixed viewBox and make the font-size
one, because SVG can't use em units in translations.

And shift the RelatedIndicator to the end of DepIndicators, because
the z index increases as you go (later entries are higher up).