DepGraph: Stub out the graph engine
Now the whole SVG is a single component.

I haven't figured out what the initialization is going to look like
yet, so we'll either need to pass the whole graph or a starting issue
into DepGraph.

This still needs automatic positioning (e.g. via graphviz), but I
wanted to get a rough outline up so I could double-check the styling
against depviz.mocks.011.jpg.

The if-true blocks address the following Firefox 45.4 warning:

  The body of a for-in should be wrapped in an if statement to filter
  unwanted properties from the prototype guard-for-in

The keys in the card array are used to identify changed entries [1].

[1]: https://facebook.github.io/react/docs/lists-and-keys.html#keys