DepGraph: Use viz.js and xml2js to position cards
This is a bit of a hack (besides the new FIXMEs in the content), since
we don't pass accurate card-size information through to vis.js (we'll
need to get at least a bit better at that to show expanded cards).

And pulling coordinates out of the graphviz SVG is currently fairly
brittle.  It would be better to walk the tree looking for the href
attributes we setup in the graphviz source.