DepCard: Add a slug property
To distinguish between the title (e.g. "single page rendering") and
the identifier (e.g. "github/jbenet/depviz#1").  Titles will usually
be longer than slugs, and we'll want to show the slug in collapsed
cards.  We don't need to show the host portion of the slug, because
the cards have a host icon.

Also fill in tests for other hosts, and throw exceptions on
unrecognized hosts.  This helps with test coverage.