GitHub: Add canonicalization support for {user}/{repo}/issues/{number}
And the /pull/ form as well.  Juan requested both of these [1,2],
presumably to allow for easier copy-pasting.  In both cases, redirect
them to the canonical-URI for the issue/pull in question.

When the user does enter one of these non-canonical keys, use an
onEnter handler to redirect them to the canonical path.

[1]: https://github.com/jbenet/depviz/issues/4
[2]: https://github.com/jbenet/depviz/issues/5