depviz.svg: Add a depviz logo
Discussion in [1].  This logo is my suggested simplicification of
Juan's DepCard-icon idea.

I also reduce the short-card relatedX offset to avoid exposing a bit
of the card border between the top and bottom DepIndicators.  That
makes the numbers more crowded as you get double-digit indicators, but
they aren't too bad.

[1]: https://github.com/jbenet/depviz/issues/14