GetNodes: Shuffle around to return an array of nodes
This lets us use initial slugs that may map to multiple nodes (e.g. a
per-repo slug) instead of the old keys which had to be one-to-one with
nodes (e.g. a per-issue slug).

Also some test shenanigans to get everything passing with 100%
coverage.  For reasons I don't understand, my attempt at atomic state
pivots are still not sticking, which is why we still can't throw an
error in the duplicate-dummy-requests-for-a-single-key case.  I'm
suspicious of [1].  But a change at duplicate requests isn't the end
of the world either.

[1]: https://github.com/facebook/react/issues/6895