DepCard: Push the border rect beneath the icon/text
I'd shifted this in 25fadc4b (DepCard: Add tasks, tasksCompleted, and
a progress bar, 2016-11-28, #18) with some rough thoughts about
clipping long slugs so they don't overflow the card.  But it meant
that the border rect masked the links provided by the icon and text.
This commit shifts the border rect lower in the stack so you can get a
click through to the active elements again.  We'll need a more robust
solution to clipping/overflow anyway.