DepCard: Expanded cards
Expanded-ness is controlled via a query parameter
(e.g. ?expanded=true), but there's no UI to toggle it at the moment
(you have to edit your location bar).  The new DepGraphView.getNodes
tunnels those settings through to the node-getters, since the DepGraph
and Graph components don't care if the nodes are expanded or not.

Falling back to the author when there are no assignees makes the
meaning of the displayed person less clear.  But it seemed more useful
to have an interested person on unassigned cards than to always stick
to assignees.  Once we get a better config UI this could be one thing
to toggle.

The comment icon is U+1F5EA (TWO SPEECH BUBBLES) [1] which is
supported by a few fonts [2], including the free-for-any-use Symbola
[3].  The task icon is U+2611 (BALLOT BOX WITH CHECK) [4] which is
widely supported [5].  The default avatar is U+1F610 (NEUTRAL FACE)
[6] with reasonable font support including Symbola [7].

[1]: http://www.fileformat.info/info/unicode/char/1f5ea/index.htm
[2]: http://www.fileformat.info/info/unicode/char/1f5ea/fontsupport.htm
[3]: http://users.teilar.gr/~g1951d/
     (at least since Symbola 8.0)
[4]: http://www.fileformat.info/info/unicode/char/2611/index.htm
[5]: http://www.fileformat.info/info/unicode/char/2611/fontsupport.htm
[6]: http://www.fileformat.info/info/unicode/char/1f610/index.htm
[7]: http://www.fileformat.info/info/unicode/char/1f610/fontsupport.htm

fixup! DepCard: Expanded cards