App: Toggle expanded/collapsed on keypress
Using onKeyPress [1].

This is more awkward than it should be, because while SVG Tiny 1.2
defines 'focusable' [2] and the draft SVG 2 defines 'tabindex' [3,4],
neither has widespread support.  In their absence, we need a wrapping
div (which does have widespread tabindex support, as a global
attribute since HTML5 [5]).  And the autofocus attribute seems to be
form-inputs-only [6,7], so we can't use that either.

Because the node expansion is tied to the node-getter, this is also
currently re-fetching and repositioning the whole graph.  That's not
very efficient.  We may want to reposition the graph to account for
the larger nodes (although we may be able to stretch collapsed
positioning), but we certainly don't want to re-fetch the existing
node data.  However, this implementation works, and an ugly, working
implementation is better than no implemementation at all ;).

[1]: https://facebook.github.io/react/docs/events.html#keyboard-events
[2]: https://www.w3.org/TR/SVGTiny12/interact.html#focusable-attr
[3]: https://svgwg.org/svg2-draft/struct.html#tabindexattribute
[4]: https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/tabindex
[5]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/tabindex
[6]: https://www.w3.org/TR/html5/forms.html#autofocusing-a-form-control:-the-autofocus-attribute
[7]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#attr-autofocus