.travis.yml: Automatically deploy master to GitHub Pages
Travis docs on encrypting settings are in [1].  Juan created a
repo-specific user [2] to work around GitHub's coarse access scoping.
That user is a collaborator on the depviz repo and created a token
with public_repo scope [3].  Juan encrypted that token with:

  $ gem install travis
  $ travis encrypt ${TOKEN}

to get the 'secure: CREv...' bit we're using in this commit.  The
deploy entry invokes the GitHub Pages deploy provider [4,5], which is
new in October [6].  The bulk of the deploy entry comes from the
testbed repo for that PR [7].  Docs for 'on:' are in [8].

[1]: https://docs.travis-ci.com/user/encryption-keys
[2]: https://github.com/depviz-ac-user
[3]: https://github.com/settings/tokens
[4]: https://docs.travis-ci.com/user/deployment/#Adding-a-Provider
[5]: https://github.com/travis-ci/dpl#github-pages
[6]: https://github.com/travis-ci/dpl/pull/501
[7]: https://github.com/webknjaz/dpl-gh-pages/blob/53fd2250c891a84e3886ed95cd685a865017b102/.travis.yml
[8]: https://docs.travis-ci.com/user/deployment/#Conditional-Releases-with-on%3