DepCard: Clip long entries in the left column
Expanded cards have three rows and potentially two columns.  Setup two
clipPath rectangles per card [1], one for the left column and one for
the whole card.  For each expandable line element starting in the left
column, associate it with:

* the left-column clipPath if that line has a right-column entry
  (e.g. keep the ID to the left column if there is a person entry to
  its right).
* the full-width clipPath if that line does not have a right-column
  entry (e.g. allow the ID to use the full width if there is no person
  entry to its right).

The new ID method ensures a valid charset for our card IDs, since SVG
1.1 punts to XML 1.0 for 'id' [2], XML 1.0 requires a Name [3], Names
have a NameStartChar and some number of NameChars [4], and those
Name*Chars match my regexps (and also contain additional non-ASCII
characters which I don't see a need to preserve yet) [5,6].

SVG 1.1 prefers to have referenced elements in a leading <defs> block
for understandability and accessibility [7].  We don't do that here
because it doesn't play nicely with the opaque-component approach that
React prefers.

Also add a title to the label list, so folks can get the full list in
a tooltip if the displayed list is clipped.

Entries in the right column and dep-indicators are usually short
enough that they don't need clipping.  The exception to that might be
people, but we can address that in later work.

[1]: https://www.w3.org/TR/SVG/masking.html#EstablishingANewClippingPath
[2]: https://www.w3.org/TR/SVG/struct.html#IDAttribute
[3]: https://www.w3.org/TR/2008/REC-xml-20081126/#id
[4]: https://www.w3.org/TR/2008/REC-xml-20081126/#NT-Name
[5]: https://www.w3.org/TR/2008/REC-xml-20081126/#NT-NameStartChar
[6]: https://www.w3.org/TR/2008/REC-xml-20081126/#NT-NameChar
[7]: https://www.w3.org/TR/SVG/struct.html#DefsElement