webapp/src/Graph.js: O(n^2) -> O(n) search for edges
We don't have to visit all node-node pairs, because we wrote the dot
source ourselves and only added edges for parents.  We'll need to
update this to also include related-issue edges once we add support
for related issues, but for the 500-odd issues in
github.com/ipfs/go-ipfs this shaves a second or so off the positioning
time.