WIP: Bookmarklet: Open depviz for the issue you're currently viewing
This still needs test coverage for run().

I'd like to have "The depviz bookmarklet" as the link text (and not
just "depviz"), but both Chromium and Firefox use the link text as the
default bookmark name.  By using only "depviz", we get a sane bookmark
name without the user having to edit anything when creating the
bookmark.

This is made tricky by GitHub's Content Security Policy, which doesn't
play nicely with all browsers [1,2,3,4].  Despite the open Chromium
issue [3], the bookmarklet does work for me with Chromium 54.0.  And
as expected from the open Firefox issue [2], the bookmarklet does not
work for me on Firefox 45.4.

[1]: https://github.com/blog/1477-content-security-policy#bookmarklets
[2]: https://bugzilla.mozilla.org/show_bug.cgi?id=866522
     Bookmarklets affected by CSP
[3]: https://bugs.chromium.org/p/chromium/issues/detail?id=233903
     CSP: Bookmarklets should bypass pages' policies.
[4]: https://www.w3.org/Bugs/Public/show_bug.cgi?id=23357
     Subverting CSP policies for browser add-ons (extensions).