DepIndicators: Click on DependenciesIndicator to load dependencies
We currently don't walk dependencies for closed issues, because
they're not related to future progress.  However, the completed-issue
graph is useful for historical context, and Juan is interested in
walking that graph [1].

To make this straightforward, we probably want a way to configure a
hop-depth from a current node.  But the current UI lacks a "current
node" concept and we haven't settled on a UI for selecting nodes
[2,3].  In the meantime, this patch gives users a way for users to
drill down into the dependencies of any closed issue by clicking on
its dependencies indicator.

unwalkedDependencies is currently a bit sloppy, since it should be
false if all of the parents have already been walked.  The current
implementation is good enough for a first pass though.

[1]: https://github.com/jbenet/depviz/issues/13#issuecomment-262672545
[2]: https://github.com/jbenet/depviz/issues/30
[3]: https://github.com/jbenet/depviz/issues/45