DepCard: Display status icon (e.g. for continuous integration tests)
Using a separate call to the GitHub API [1] for open pull requests,
because GitHub doesn't expose this information directly in its issue
response.  The hack access to _request works around the
currently-missing github-tools/github wrapper for this API endpoint
[2].

We don't look this up for closed pull requests, because a separate
per-issue request is expensive, and users are unlikely to care that
much about the CI status of closed pull requests.  If it turns out
that some do care, we'll have to make the done-ness filter
configurable.

This commit also shuffles the expanded card from a two-column,
three-line layout to a one-column, four-line layout (although the last
line is split into a number of columns).  That gives us more space for
multiple people and longer IDs/titles, keeps the cards at a reasonable
size, and doesn't crowd too much.

The status icons are:

* success: U+2713 (CHECK MARK).
* failure: U+26A0 (WARNING SIGN).  Also considered: U+2717 (BALLOT X).
* pending: U+231B (HOURGLASS).  Also considered:
  * U+25CF (BLACK CIRCLE)
  * U+2022 (BULLET)
  * U+23F2 (TIMER CLOCK)

[1]: https://developer.github.com/v3/repos/statuses/#get-the-combined-status-for-a-specific-ref
[2]: https://github.com/github-tools/github/pull/397